requirements.txt
selenium==4.25.0
python-dotenv==1.0.1
pymongo==4.10.1
requests==2.32.3
pandas==2.3.3